Fix broken URLs
Here are some propositions regarding the broken URLs since the official project has been stopped. This changes the URL of the dnscrypt-proxy .tar.gz to the Ubuntu archive repo, hardcoded to version 1.9.4, adds a hardcoded signature to verify the .tar.gz and replaces the resolvers .csv URL to Dyne's version in their GitHub repo. I have tested this to be working correctly in a Debian 9.3.0 live environment.